refactor(e2e): split baseline conftest into lane_selection + _fixtures
conftest.py was 418 lines mixing pytest hooks, CI-lane logic, and ~250 lines
of fixtures. Separate the concerns so each file does one job:

- lane_selection.py: the BAND_E2E_LANE collection logic
  (_item_target_adapters, _lane_skip_reason, apply_lane_skips); the conftest
  hook is now a one-line delegate.
- _fixtures/{platform,agents,capture}.py: fixtures grouped by concern,
  re-imported into conftest. pytest_plugins is deprecated in a non-root
  conftest, so the import-re-export keeps fixtures registered and scoped to
  the baseline subtree (listed in __all__ so they read as intentional).

conftest.py drops to hooks + re-exports (~115 lines). No behavior change:
76 tests still collected, lane scoping intact, registry guards pass.
