+++ tests/e2e/baseline/conftest.py
@@ -1,49 +1,54 @@
-"""Fixtures for the baseline testing toolkit.
+"""Pytest wiring for the baseline toolkit.
 
-Config comes from the concern-separated ``BaselineSettings`` (see settings.py),
-not the legacy flat ``E2ESettings``. Provisioning (provision/reap) and the other
-tools add their fixtures here as they are built.
+This file holds only the pytest *glue*: the marker registration, the always-on
+E2E/Band-key gate, and the CI-lane collection hook. The fixtures themselves live
+in ``_fixtures/`` (platform / agents / capture) and the lane-selection logic in
+``lane_selection``; both are imported here so pytest registers the fixtures for
+the baseline subtree (``pytest_plugins`` is deprecated in a non-root conftest).
+The fixture re-exports are listed in ``__all__`` so they read as intentional.
 """
 
 from __future__ import annotations
 
-import functools
-import os
-from collections.abc import AsyncGenerator, Awaitable, Callable
-from contextlib import AbstractAsyncContextManager, AsyncExitStack
-
 import pytest
-from anthropic import AsyncAnthropic
-from band_rest import AsyncRestClient
 
-from band.client.streaming import WebSocketClient
-
-from tests.e2e.baseline.agents import (
-    AGENTS_MARKER,
-    MATRIX_MARKER,
-    AgentsRequest,
-    MatrixBuild,
+from tests.e2e.baseline._fixtures.agents import (
+    adapter_id,
+    agent,
+    agents,
+    matrix_agent,
 )
-from tests.e2e.baseline.requires import MARKER, Dep, require_dep, requires
+from tests.e2e.baseline._fixtures.capture import judge, reply_capture
+from tests.e2e.baseline._fixtures.platform import (
+    baseline_run_id,
+    baseline_settings,
+    baseline_user_client,
+    baseline_ws,
+    orphan_sweep,
+    resource_manager,
+    user_ops,
+)
+from tests.e2e.baseline.agents import AGENTS_MARKER, MATRIX_MARKER
+from tests.e2e.baseline.lane_selection import apply_lane_skips
+from tests.e2e.baseline.requires import MARKER, require_dep
 from tests.e2e.baseline.settings import BaselineSettings
-from tests.e2e.baseline.toolkit.adapters import (
-    build_adapter,
-    ci_lanes,
-    infra_adapters,
-    specs,
-)
-from tests.e2e.baseline.toolkit.judge import Verdict
-from tests.e2e.baseline.toolkit.judge import judge as _judge
-from tests.e2e.baseline.toolkit.provisioning import (
-    ProvisionedAgent,
-    ResourceManager,
-    new_run_id,
-    running_provisioned_agent,
-)
-from tests.e2e.baseline.toolkit.user_ops import UserOps
-from tests.e2e.baseline.toolkit.capture import ReplyCapture
-from tests.e2e.baseline.toolkit.capture import reply_capture as _reply_capture
-from tests.e2e.helpers import TrackingWebSocketClient
+
+# Re-exported fixtures (defined in _fixtures/*; imported so pytest registers them).
+__all__ = [
+    "adapter_id",
+    "agent",
+    "agents",
+    "baseline_run_id",
+    "baseline_settings",
+    "baseline_user_client",
+    "baseline_ws",
+    "judge",
+    "matrix_agent",
+    "orphan_sweep",
+    "reply_capture",
+    "resource_manager",
+    "user_ops",
+]
 
 
 def pytest_configure(config: pytest.Config) -> None:
@@ -86,333 +91,8 @@ def pytest_runtest_setup(item: pytest.Item) -> None:
             require_dep(dep, settings)
 
 
-def _item_target_adapters(item: pytest.Item) -> frozenset[str]:
-    """The adapter ids an item is bound to (empty = adapter-agnostic, always kept).
-
-    A matrix cell carries its id as the ``adapter_id`` callspec param; a
-    ``@with_agents`` test carries its adapters on the AGENTS_MARKER. Tests with
-    neither (provisioning, user-ops, the registry guard) target no adapter and run
-    in every lane.
-    """
-    callspec = getattr(item, "callspec", None)
-    if callspec is not None and "adapter_id" in callspec.params:
-        return frozenset({str(callspec.params["adapter_id"])})
-    marker = item.get_closest_marker(AGENTS_MARKER)
-    if marker is not None:
-        request: AgentsRequest = marker.args[0]
-        return frozenset(str(adapter) for adapter in request.adapters)
-    return frozenset()
-
-
-def _lane_skip_reason(
-    targets: frozenset[str],
-    lane: str,
-    lane_of: dict[str, str],
-    infra: frozenset[str],
-) -> str | None:
-    """Why ``targets`` can't run in ``lane`` (skip reason), or ``None`` if in-lane.
-
-    Skip — never fail — for the two deliberate out-of-scope cases: an infra adapter
-    (external backend not wired in CI) or an adapter belonging to a different lane's
-    venv. An *in-lane* adapter is left untouched so its ``@requires`` gate still
-    *fails* on a missing provider key (misconfig stays loud). ``lane_of`` maps each
-    CI-runnable adapter to its lane (from ``ci_lanes``).
-    """
-    infra_hit = sorted(targets & infra)
-    if infra_hit:
-        return f"{infra_hit} need an external backend; no CI lane runs them yet"
-    out_of_lane = sorted(t for t in targets if lane_of.get(t) != lane)
-    if out_of_lane:
-        elsewhere = sorted({lane_of.get(t, "?") for t in out_of_lane})
-        return f"{out_of_lane} run in lane(s) {elsewhere}, not active lane {lane!r}"
-    return None
-
-
 def pytest_collection_modifyitems(
     config: pytest.Config, items: list[pytest.Item]
 ) -> None:
-    """Scope the run to ``BAND_E2E_LANE`` (a ``uv`` extra), resolved via the registry.
-
-    The lane's adapter set is *derived* (``ci_lanes`` / ``infra_adapters``), never a
-    hand list — so a newly-registered adapter joins its lane with no change here.
-    Every test bound to an out-of-lane or infra adapter is marked **skip-with-reason**
-    (visible in the lane's report); in-lane tests are untouched, so a missing
-    provider key still fails via the ``@requires`` gate. Adapter-agnostic tests
-    always run. Unset ``BAND_E2E_LANE`` leaves collection untouched (full matrix,
-    fail-loud) — the correct local default.
-    """
-    lane = os.environ.get("BAND_E2E_LANE")
-    if lane is None:
-        return
-    lanes = ci_lanes()
-    if lane not in lanes:
-        raise pytest.UsageError(
-            f"BAND_E2E_LANE={lane!r} is not a known CI lane; registry lanes are "
-            f"{sorted(lanes)}"
-        )
-    lane_of = {str(a): extra for extra, ids in lanes.items() for a in ids}
-    infra = frozenset(str(a) for a in infra_adapters())
-    for item in items:
-        targets = _item_target_adapters(item)
-        if not targets:
-            continue
-        reason = _lane_skip_reason(targets, lane, lane_of, infra)
-        if reason is not None:
-            item.add_marker(pytest.mark.skip(reason=reason))
-
-
-@pytest.fixture(scope="session")
-def baseline_settings() -> BaselineSettings:
-    return BaselineSettings()
-
-
-@pytest.fixture(scope="session")
-def baseline_run_id() -> str:
-    """Token identifying this session's provisioned resources (for naming/sweep)."""
-    return new_run_id()
-
-
-@pytest.fixture(scope="session")
-def baseline_user_client(baseline_settings: BaselineSettings) -> AsyncRestClient:
-    """Session-scoped user-authenticated REST client for provisioning."""
-    assert baseline_settings.credentials.api_key_user, (
-        "BAND_API_KEY_USER is required for provisioning"
-    )
-    return AsyncRestClient(
-        api_key=baseline_settings.credentials.api_key_user,
-        base_url=baseline_settings.endpoints.rest_url,
-    )
-
-
-@pytest.fixture
-def user_ops(baseline_user_client: AsyncRestClient) -> UserOps:
-    """User-operation driver over the session-scoped user REST client.
-
-    Reuses ``baseline_user_client`` (which already requires BAND_API_KEY_USER)
-    rather than spinning up a fresh client per test.
-    """
-    return UserOps(baseline_user_client)
-
-
-@pytest.fixture
-async def judge(
-    baseline_settings: BaselineSettings,
-) -> AsyncGenerator[Callable[..., Awaitable[Verdict]], None]:
-    """LLM judge with the client + model pre-bound; call with criteria/transcript.
-
-    Self-gates on its provider key so any test using it skips cleanly when the
-    key is absent — the requirement travels with the fixture. The Anthropic
-    client is built once here (and closed on teardown) rather than per verdict.
-
-    Usage::
-
-        verdict = await judge(criteria="...", transcript="...")
-    """
-    require_dep(Dep.ANTHROPIC, baseline_settings)
-    async with AsyncAnthropic(
-        api_key=baseline_settings.llm_credentials.anthropic_api_key
-    ) as client:
-        yield functools.partial(
-            _judge, client=client, model=baseline_settings.llm_models.judge_model
-        )
-
-
-@pytest.fixture(scope="session")
-async def baseline_ws(
-    baseline_settings: BaselineSettings,
-) -> AsyncGenerator[TrackingWebSocketClient, None]:
-    """User-authenticated WS observer for the wait primitives.
-
-    Connects as the user (not an agent), so it coexists with agents and
-    receives the same ``message_created`` events. Session-scoped to avoid
-    per-test connect/teardown latency; channels are left on teardown.
-    """
-    assert baseline_settings.credentials.api_key_user, (
-        "BAND_API_KEY_USER is required for the WS observer"
-    )
-    ws = WebSocketClient(
-        ws_url=baseline_settings.endpoints.ws_url,
-        api_key=baseline_settings.credentials.api_key_user,
-        agent_id=None,  # user connection, not an agent
-    )
-    async with ws:
-        tracking = TrackingWebSocketClient(ws)
-        yield tracking
-        await tracking.cleanup_channels()
-
-
-# Shared agent prompt for the demo adapters: short replies via the tool.
-_SHORT_PROMPT = (
-    "Keep responses to one short sentence. Always reply using band_send_message."
-)
-
-
-@pytest.fixture
-async def agents(
-    request: pytest.FixtureRequest,
-    baseline_settings: BaselineSettings,
-    resource_manager: ResourceManager,
-) -> AsyncGenerator[list[ProvisionedAgent], None]:
-    """The running agents declared by ``@with_agents(...)``, in declared order.
-
-    Builds each adapter via the registry (steered by the decorator's
-    ``prompt``/``features``, defaulting to the short prompt), runs each through
-    ``running_provisioned_agent`` (reaped by ``resource_manager`` teardown), and
-    yields the ``ProvisionedAgent`` records. Use ``agent`` for the single case.
-
-    Each slot gets an index-suffixed label so the same framework can appear more
-    than once (e.g. ``@with_agents(Adapter.ANTHROPIC, Adapter.ANTHROPIC)`` for two
-    same-type agents in one room) without provisioned-name collisions.
-    """
-    marker = request.node.get_closest_marker(AGENTS_MARKER)
-    if marker is None:
-        raise pytest.UsageError(
-            "the `agents`/`agent` fixture requires the @with_agents(...) decorator"
-        )
-    req: AgentsRequest = marker.args[0]
-    prompt = req.prompt if req.prompt is not None else _SHORT_PROMPT
-    async with AsyncExitStack() as stack:
-        provisioned = [
-            await stack.enter_async_context(
-                running_provisioned_agent(
-                    build_adapter(
-                        name,
-                        baseline_settings,
-                        prompt=prompt,
-                        features=req.features,
-                        tools=req.tools,
-                    ),
-                    resource_manager,
-                    label=f"{name}-{slot}",
-                )
-            )
-            for slot, name in enumerate(req.adapters)
-        ]
-        yield provisioned
-
-
-@pytest.fixture
-async def agent(agents: list[ProvisionedAgent]) -> ProvisionedAgent:
-    """The single running agent declared by ``@with_agents(OneAdapter)``."""
-    if len(agents) != 1:
-        raise pytest.UsageError(
-            f"`agent` needs exactly one adapter in @with_agents(...); got "
-            f"{len(agents)} — use `agents` for multiple."
-        )
-    return agents[0]
-
-
-# The full adapter matrix as fixture params: one cell per registered adapter, each
-# carrying its ``@requires`` marks so a missing requirement fails that cell. Built
-# once at import (registration happens when ``toolkit.adapters`` is imported above).
-_MATRIX_PARAMS = [
-    pytest.param(spec.id, marks=requires(*spec.requires), id=str(spec.id))
-    for spec in specs()
-]
-
-
-@pytest.fixture(params=_MATRIX_PARAMS)
-def adapter_id(request: pytest.FixtureRequest) -> str:
-    """The current matrix cell's adapter id.
-
-    Parametrized across the whole registry by default (each cell carrying its
-    ``@requires`` gate), so requesting this — or ``matrix_agent``, which depends on
-    it — fans a test across the matrix. ``@across_adapters(...)`` overrides the set;
-    construction-only tests parametrize it directly.
-    """
-    return request.param
-
-
-@pytest.fixture
-async def matrix_agent(
-    request: pytest.FixtureRequest,
-    adapter_id: str,
-    baseline_settings: BaselineSettings,
-    resource_manager: ResourceManager,
-) -> AsyncGenerator[ProvisionedAgent, None]:
-    """The running provisioned agent for the current matrix cell.
-
-    Built from ``adapter_id`` via the registry and run for the test; request
-    ``adapter_id`` alongside it when a test also needs the id. ``@across_adapters``
-    may steer construction via the ``MatrixBuild`` marker (``prompt`` / ``features``
-    — e.g. enable memory), else a short default prompt and no features. Reaping is
-    owned by ``resource_manager`` teardown.
-    """
-    marker = request.node.get_closest_marker(MATRIX_MARKER)
-    build: MatrixBuild | None = marker.args[0] if marker else None
-    prompt = build.prompt if build and build.prompt is not None else _SHORT_PROMPT
-    features = build.features if build else None
-    tools = build.tools if build else None
-    adapter = build_adapter(
-        adapter_id, baseline_settings, prompt=prompt, features=features, tools=tools
-    )
-    async with running_provisioned_agent(
-        adapter, resource_manager, label=adapter_id
-    ) as provisioned:
-        yield provisioned
-
-
-@pytest.fixture
-def reply_capture(
-    baseline_ws: TrackingWebSocketClient,
-    baseline_settings: BaselineSettings,
-    user_ops: UserOps,
-) -> Callable[[str], AbstractAsyncContextManager[ReplyCapture]]:
-    """Subscribe-before-send capture with the WS observer + E2E_TIMEOUT pre-bound.
-
-    Hides ``baseline_ws`` from tests; use as ``async with reply_capture(room_id)``.
-    The capture's default wait deadline comes from E2E_TIMEOUT. ``user_ops`` and
-    ``settings`` are pre-bound so ``capture.tool_calls()`` / ``events()`` /
-    ``memory(agent)`` can read persisted events and agent-scoped memory.
-    """
-    return functools.partial(
-        _reply_capture,
-        baseline_ws,
-        user_ops=user_ops,
-        settings=baseline_settings,
-        deadline_s=baseline_settings.e2e_timeout,
-    )
-
-
-@pytest.fixture
-async def resource_manager(
-    baseline_settings: BaselineSettings,
-    baseline_user_client: AsyncRestClient,
-    baseline_run_id: str,
-) -> AsyncGenerator[ResourceManager, None]:
-    """Per-test provision/reap driver.
-
-    Teardown force-deletes everything provisioned this run, unless
-    ``BAND_E2E_AUTOCLEAN`` is false (kept for on-purpose debugging; surviving
-    ids are logged by ``reap_all``/``provision_*``).
-    """
-    resources = ResourceManager(
-        user_client=baseline_user_client,
-        settings=baseline_settings,
-        run_id=baseline_run_id,
-    )
-    yield resources
-    if baseline_settings.run.autoclean:
-        await resources.reap_all()
-
-
-@pytest.fixture(scope="session", autouse=True)
-async def orphan_sweep(
-    baseline_settings: BaselineSettings,
-    baseline_user_client: AsyncRestClient,
-    baseline_run_id: str,
-) -> None:
-    """Reap stale test agents from crashed prior runs, once at session start.
-
-    Prefix-guarded and age-guarded (see ``ResourceManager.sweep_orphans``), so
-    it never deletes a non-test agent or a concurrent run's fresh resources.
-    No-op when ``BAND_E2E_ORPHAN_SWEEP`` is false.
-    """
-    if not baseline_settings.run.orphan_sweep:
-        return
-    resources = ResourceManager(
-        user_client=baseline_user_client,
-        settings=baseline_settings,
-        run_id=baseline_run_id,
-    )
-    await resources.sweep_orphans()
+    """Scope the run to ``BAND_E2E_LANE`` (see ``lane_selection``)."""
+    apply_lane_skips(config, items)
